Revert jwt names to ensure compatibility
