+++ SFA.DAS.Http/SFA.DAS.Http/Configuration/IGenericJwtClientConfiguration.cs
@@ -1,0 +1,11 @@
+﻿namespace SFA.DAS.Http.Configuration
+{
+    public interface IGenericJwtClientConfiguration
+    {
+        // JWT Configuration
+        string Issuer {get; }
+        string Audience {get; }
+        string ClientSecret {get; }
+        int TokenExpirySeconds {get; }
+    }
+}
