--- SFA.DAS.Http/SFA.DAS.Http/Configuration/IStaticJwtClientConfiguration.cs
@@ -1,8 +1,0 @@
-﻿namespace SFA.DAS.Http.Configuration
-{
-    public interface IStaticJwtClientConfiguration
-    {
-        // JWT Configuration
-        string ClientToken { get; }
-    }
-}
